Some cleanup.

+++ src/MainMenu.h
@@ -29,7 +29,7 @@ struct MainMenu
    struct SDL_Surface* m_pScreen;//Does not own
    struct SDL_Surface* m_pFlagsSurface;
    struct FlagInformation* m_pFlagInformation;
-   short* m_pbAlreadyLoadedImages;
+   int* m_pbAlreadyLoadedImages;
 };
 
 void CreateMainMenu(struct MainMenu** ppMenu, struct Config* pConfig, struct SDL_Surface* pScreen);
